+++ CodeChat/CodeToRestSphinx.py
@@ -37,6 +37,7 @@
 #
 # Standard library
 # ----------------
+import re
 from os import path
 import fnmatch
 import codecs
@@ -84,8 +85,14 @@ def _source_read(
             # this will raise an exception on failure.
             lexer = lexer or get_lexer(filename=docname, code=source[0])
 
+            # Translate code to reST.
             app.info('Converted using the {} lexer.'.format(lexer.name))
             source[0] = code_to_rest_string(source[0], lexer=lexer)
+
+            # Add in the highlight language to use, unless there's potentially `file-wide metadata <http://www.sphinx-doc.org/en/stable/markup/misc.html#file-wide-metadata>`_. It's hard to know in this case where the ``.. highlight`` directive can be safely placed. Putting it before file-wide metadata demotes it to not being metadata. Finding the right place to put the ``.. highlight`` directive after the metadata is difficult to know.
+            if not re.search('^:(tocdepth|nocomments|orphan):', source[0], re.MULTILINE):
+                source[0] = '.. highlight:: {}\n\n{}'.format(lexer.name, source[0])
+
         except (KeyError, pygments.util.ClassNotFound) as e:
             # We don't support this language.
             app.warn(e, docname)