+++ docs/history.rst
@@ -16,9 +16,13 @@ History of recent changes
 
     -   No changes yet.
 
+-   1.5.5, 18-Oct-2017
+
+    -   Automatically insert a ``.. highlight`` directive based on the lexer used, unless the file contains `file-wide metadata <http://www.sphinx-doc.org/en/stable/markup/misc.html#file-wide-metadata>`_.
+
 -   1.5.4, 29-Sep-2017
 
-    -   Updated template CSS to latest.
+    -   Update template CSS to latest.
 
 -   1.5.3, 29-Sep-2017